Typescript trackhubregistry
